Prefer to use Travis-CI's rebar, test more OTP versions.
Unless the bundled version of the rebar binary is built with the least supported
version of OTP in .travis.yml it is possible that op-code not backwards
compatible will sneak in and cause test build errors. Alternatively, one may
choose to use travis-ci's rebar binary.

This commit removes the bundled rebar binary in favor of one found on the path.
This requires that developers have a rebar binary in their PATH.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>